fix old storage crash

+++ app/src/main/java/com/github/axet/bookreader/app/Storage.java
@@ -920,6 +920,7 @@ public class Storage extends com.github.axet.androidlibrary.app.Storage {
     }
 
     public Book load(InputStream is, File f) {
+        boolean tmp = f == null;
         Book fbook = new Book();
         try {
             FileOutputStream os = null;
@@ -962,15 +963,15 @@ public class Storage extends com.github.axet.androidlibrary.app.Storage {
                     if (d instanceof FileTypeDetectorZipExtract.Handler) {
                         FileTypeDetectorZipExtract.Handler e = (FileTypeDetectorZipExtract.Handler) d;
                         File z = fbook.file;
-                        if (f != null) {
-                            f = null; // force to delete
-                            fbook.file = File.createTempFile("book", ".tmp");
+                        if (!tmp) {
+                            fbook.file = File.createTempFile("book", ".tmp", context.getCacheDir());
                             fbook.md5 = e.extract(z, fbook.file);
+                            tmp = true; // force to delete 'fbook.file'
                         } else {
-                            File tmp = File.createTempFile("book", ".tmp");
-                            fbook.md5 = e.extract(z, tmp);
+                            File tt = File.createTempFile("book", ".tmp", context.getCacheDir());
+                            fbook.md5 = e.extract(z, tt);
                             z.delete();
-                            fbook.file = tmp;
+                            fbook.file = tt;
                         }
                     }
                     break; // priority first - more imporant
@@ -983,12 +984,12 @@ public class Storage extends com.github.axet.androidlibrary.app.Storage {
         } catch (RuntimeException e) {
             throw e;
         } catch (IOException | NoSuchAlgorithmException e) {
-            if (fbook.file != null)
+            if (tmp && fbook.file != null)
                 fbook.file.delete();
             throw new RuntimeException(e);
         }
         File storage = getLocalStorage();
-        fbook.store(storage, f == null);
+        fbook.store(storage, tmp);
         return fbook;
     }
